Setup connection to Heroku and Docker
The readme has a complete tutorial with documentation for how to clone, setup, and run the application.

Everyone will have their own docker container on their system. As long as we all have updated code when we start working and we check the project runs in the docker image before pushing, we should be good.